o Added a new generic LdapException o Wrapped all the exception to produce LdapException o Finished the BindRequest implementation o Commented the other operation atm
git-svn-id: https://svn.apache.org/repos/asf/directory/shared/branches/shared-replication@760567 13f79535-47bb-0310-9956-ffa450edef68

+++ client-api/src/main/java/org/apache/directory/shared/ldap/client/api/messages/BindRequestImpl.java
@@ -56,7 +56,7 @@ public class BindRequestImpl implements BindRequest
     private Map<String, Control> controls;
     
     /** The client request timeout */
-    private int timeout = 0;
+    private long timeout = 0;
 
 
     /**
@@ -209,7 +209,7 @@ public class BindRequestImpl implements BindRequest
     /**
      * {@inheritDoc}
      */
-    public int getTimeout()
+    public long getTimeout()
     {
         return timeout;
     }
@@ -239,7 +239,7 @@ public class BindRequestImpl implements BindRequest
     /**
      * {@inheritDoc}
      */
-    public void setTimeout( int timeout )
+    public void setTimeout( long timeout )
     {
         this.timeout = timeout;
     }
